klash-backend/src/python/requirements.txt
twscrape>=0.7.0
textblob>=0.17.1
spacy>=3.7.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2
transformers>=4.30.0
torch>=2.0.0
tqdm>=4.65.0
numpy>=1.24.0
